Fix template command to not print list of deleted files
For all wrapped Helm commands the list of decrypted files is printed as
they get removed at the end of plugin execution. For the `template`
command this is undesired, as it means you cannot simply run `helm
secrets template` and get valid YAML output.

This change suppresses the output of deleted files for the `template`
only, meaning it does not alter the behavior of other commands, but just
fixes the `template` command to work as I imagine most people would
expect it to work.

[Fixes issues #126 and #132]